Add service to remove record and segment IDs
The output of the `convertToValues` service contains metadata about
the internal processing. Those IDs are usually not needed for
processing the payload. They can be removed using the service
`jt.flatFile.pub.read:removeRecordAndSegmentIds` .

The sample Flow service `jt.flatFile.test.pub.commaSeparated:process`
now makes use of `removeRecordAndSegmentIds` to demonstrate its
effect. The sample for semicolon as separator was not changed and
still returns the IDs for comparison.